Use new `TextPreview` type to generate rich-Html previews of all events
Used for a preview of the latest message in each room of the rooms list,
and for inline replies, both in the timeline and above the message input box
when drafting a reply to an earlier message.
