+++ src/home/room_screen.rs
@@ -6,21 +6,32 @@ use std::{borrow::Cow, collections::BTreeMap, ops::{Deref, DerefMut, Range}, syn
 use imbl::Vector;
 use makepad_widgets::*;
 use matrix_sdk::{ruma::{
-    events::{
+    events::
         room::{
-            guest_access::GuestAccess, history_visibility::HistoryVisibility, join_rules::JoinRule, message::{ImageMessageEventContent, MessageFormat, MessageType, RoomMessageEventContent, TextMessageEventContent}, MediaSource
-        },
-        AnySyncMessageLikeEvent, AnySyncTimelineEvent, FullStateEventContent, SyncMessageLikeEvent,
-    }, matrix_uri::MatrixId, uint, EventId, MatrixToUri, MatrixUri, MilliSecondsSinceUnixEpoch, OwnedEventId, OwnedRoomId, RoomId, UserId
+            message::{ImageMessageEventContent, MessageFormat, MessageType, RoomMessageEventContent, TextMessageEventContent}, MediaSource
+        }
+    , matrix_uri::MatrixId, uint, EventId, MatrixToUri, MatrixUri, MilliSecondsSinceUnixEpoch, OwnedEventId, OwnedRoomId, RoomId, UserId
 }, OwnedServerName};
 use matrix_sdk_ui::timeline::{
-    self, AnyOtherFullStateEventContent, EventTimelineItem, MemberProfileChange, MembershipChange,
-    Profile, ReactionsByKeyBySender, RepliedToInfo, RoomMembershipChange,
+    self, EventTimelineItem, MemberProfileChange, Profile, ReactionsByKeyBySender, RepliedToInfo, RoomMembershipChange,
     TimelineDetails, TimelineItem, TimelineItemContent, TimelineItemKind, VirtualTimelineItem,
 };
 
 use crate::{
-    avatar_cache::{self, AvatarCacheEntry}, media_cache::{MediaCache, MediaCacheEntry}, profile::{user_profile::{AvatarState, ShowUserProfileAction, UserProfile, UserProfileAndRoomId, UserProfilePaneInfo, UserProfileSlidingPaneRef, UserProfileSlidingPaneWidgetExt}, user_profile_cache}, shared::{avatar::{AvatarRef, AvatarWidgetRefExt}, html_or_plaintext::{HtmlOrPlaintextRef, HtmlOrPlaintextWidgetRefExt}, text_or_image::{TextOrImageRef, TextOrImageWidgetRefExt}}, sliding_sync::{get_client, submit_async_request, take_timeline_update_receiver, MatrixRequest}, utils::{self, unix_time_millis_to_datetime, MediaFormatConst}
+    avatar_cache::{self, AvatarCacheEntry},
+    event_preview::{text_preview_of_member_profile_change, text_preview_of_other_state, text_preview_of_redacted_message, text_preview_of_room_membership_change, text_preview_of_timeline_item},
+    media_cache::{MediaCache, MediaCacheEntry},
+    profile::{
+        user_profile::{AvatarState, ShowUserProfileAction, UserProfile, UserProfileAndRoomId, UserProfilePaneInfo, UserProfileSlidingPaneRef, UserProfileSlidingPaneWidgetExt},
+        user_profile_cache,
+    },
+    shared::{
+        avatar::{AvatarRef, AvatarWidgetRefExt},
+        html_or_plaintext::{HtmlOrPlaintextRef, HtmlOrPlaintextWidgetRefExt},
+        text_or_image::{TextOrImageRef, TextOrImageWidgetRefExt},
+    },
+    sliding_sync::{get_client, submit_async_request, take_timeline_update_receiver, MatrixRequest},
+    utils::{self, unix_time_millis_to_datetime, MediaFormatConst},
 };
 use rangemap::RangeSet;
 
@@ -169,7 +180,6 @@ live_design! {
         reply_preview_body = <HtmlOrPlaintext> {
             html_view = { html = {
                 font_size: (MESSAGE_REPLY_PREVIEW_FONT_SIZE)
-                font_size: (MESSAGE_REPLY_PREVIEW_FONT_SIZE),
                     draw_normal:      { wrap: Ellipsis, text_style: { font_size: (MESSAGE_REPLY_PREVIEW_FONT_SIZE) } },
                     draw_italic:      { wrap: Ellipsis, text_style: { font_size: (MESSAGE_REPLY_PREVIEW_FONT_SIZE) } },
                     draw_bold:        { wrap: Ellipsis, text_style: { font_size: (MESSAGE_REPLY_PREVIEW_FONT_SIZE) } },
@@ -2036,141 +2046,6 @@ fn draw_replied_to_message(
 }
 
 
-/// Returns a text preview of the given timeline event as an Html-formatted string.
-pub fn preview_text_of_timeline_item(
-    content: &TimelineItemContent,
-    sender_username: &str,
-) -> String {
-    match content {
-        TimelineItemContent::Message(m) => preview_text_of_message(m, sender_username),
-        TimelineItemContent::RedactedMessage => format!(
-            "[Message was redacted]",
-        ),
-        TimelineItemContent::Sticker(sticker) => format!(
-            "[Sticker]: <i>{}</i>",
-            sticker.content().body,
-        ),
-        TimelineItemContent::UnableToDecrypt(_encrypted_msg) => format!(
-            "[Unable to decrypt message]",
-        ),
-        TimelineItemContent::MembershipChange(membership_change) => {
-            preview_text_of_room_membership_change(membership_change)
-                .unwrap_or_else(|| String::from("<i>underwent a membership change</i>"))
-        }
-        TimelineItemContent::ProfileChange(profile_change) => {
-            preview_text_of_member_profile_change(profile_change, sender_username)
-        }
-        TimelineItemContent::OtherState(other_state) => {
-            preview_text_of_other_state(other_state)
-                .unwrap_or_else(|| String::from("<i>initiated another state change</i>"))
-        }
-        TimelineItemContent::FailedToParseMessageLike { event_type, .. } => format!(
-            "[Failed to parse <i>{}</i> message]",
-            event_type,
-        ),
-        TimelineItemContent::FailedToParseState { event_type, .. } => format!(
-            "[Failed to parse <i>{}</i> state]",
-            event_type,
-        ),
-        TimelineItemContent::Poll(poll_state) => format!(
-            "[Poll]: {}",
-            poll_state.fallback_text()
-                .unwrap_or_else(|| poll_state.results().question),
-        ),
-        TimelineItemContent::CallInvite => format!(
-            "[Call Invitation]",
-        ),
-        TimelineItemContent::CallNotify => format!(
-            "[Call Notification]",
-        ),
-    }
-}
-
-pub fn preview_text_of_message(
-    message: &timeline::Message,
-    sender_username: &str,
-) -> String {
-    match message.msgtype() {
-        MessageType::Audio(audio) => format!(
-            "[Audio]: <i>{}</i>",
-            if let Some(formatted_body) = audio.formatted.as_ref() {
-                &formatted_body.body
-            } else {
-                &audio.body
-            }
-        ),
-        MessageType::Emote(emote) => format!(
-            "<i>{} {}</i>",
-            sender_username,
-            if let Some(formatted_body) = emote.formatted.as_ref() {
-                &formatted_body.body
-            } else {
-                &emote.body
-            }
-        ),
-        MessageType::File(file) => format!(
-            "[File]: <i>{}</i>",
-            if let Some(formatted_body) = file.formatted.as_ref() {
-                &formatted_body.body
-            } else {
-                &file.body
-            }
-        ),
-        MessageType::Image(image) => format!(
-            "[Image]: <i>{}</i>",
-            if let Some(formatted_body) = image.formatted.as_ref() {
-                &formatted_body.body
-            } else {
-                &image.body
-            }
-        ),
-        MessageType::Location(location) => format!(
-            "[Location]: <i>{}</i>",
-            location.body,
-        ),
-        MessageType::Notice(notice) => format!("[Notice]: <i>{}</i>",
-            if let Some(formatted_body) = notice.formatted.as_ref() {
-                &formatted_body.body
-            } else {
-                &notice.body
-            }
-        ),
-        MessageType::ServerNotice(notice) => format!(
-            "[Server Notice]: <i>{} -- {}</i>",
-            notice.server_notice_type.as_str(),
-            notice.body,
-        ),
-        MessageType::Text(text) => {
-            text.formatted.as_ref()
-                .and_then(|fb| (fb.format == MessageFormat::Html)
-                    .then(|| utils::linkify(&fb.body).to_string())
-                )
-                .unwrap_or_else(|| utils::linkify(&text.body).to_string())
-        }
-        MessageType::VerificationRequest(verification) => format!(
-            "[Verification Request] <i>from device {} to user {}</i>",
-            verification.from_device,
-            verification.to,
-        ),
-        MessageType::Video(video) => format!(
-            "[Video]: <i>{}</i>",
-            if let Some(formatted_body) = video.formatted.as_ref() {
-                &formatted_body.body
-            } else {
-                &video.body
-            }
-        ),
-        MessageType::_Custom(custom) => format!(
-            "[Custom message]: {:?}",
-            custom,
-        ),
-        other => format!(
-            "[Unknown message type]: {}",
-            other.body(),
-        )
-    }
-}
-
 fn populate_preview_of_timeline_item(
     widget_out: &HtmlOrPlaintextRef,
     timeline_item_content: &TimelineItemContent,
@@ -2182,7 +2057,8 @@ fn populate_preview_of_timeline_item(
         }
     }
     widget_out.show_html(
-        preview_text_of_timeline_item(timeline_item_content, sender_username)
+        text_preview_of_timeline_item(timeline_item_content, sender_username)
+            .format_with(sender_username)
     );
 }
 
@@ -2274,51 +2150,11 @@ impl SmallStateEventContent for RedactedMessageEventMarker {
         mut new_drawn_status: ItemDrawnStatus,
     ) -> (WidgetRef, ItemDrawnStatus) {
         item.label(id!(content)).set_text(
-            &preview_text_of_redacted_message(event_tl_item, original_sender)
+            &text_preview_of_redacted_message(event_tl_item, original_sender)
+                .format_with(original_sender)
         );
         new_drawn_status.content_drawn = true;
         (item, new_drawn_status)
-    }
-}
-
-pub fn preview_text_of_redacted_message(
-    event_tl_item: &EventTimelineItem,
-    original_sender: &str,
-) -> String {
-    let redactor_and_reason = {
-        let mut rr = None;
-        if let Some(redacted_msg) = event_tl_item.latest_json() {
-            if let Ok(old) = redacted_msg.deserialize() {
-                if let AnySyncTimelineEvent::MessageLike(
-                    AnySyncMessageLikeEvent::RoomMessage(
-                        SyncMessageLikeEvent::Redacted(redaction)
-                    )
-                ) = old {
-                    rr = Some((
-                        redaction.unsigned.redacted_because.sender,
-                        redaction.unsigned.redacted_because.content.reason,
-                    ));
-                }
-            }
-        }
-        rr
-    };
-
-    match redactor_and_reason {
-        Some((redactor, Some(reason))) => {
-            // TODO: get the redactor's display name if possible
-            format!("{} deleted {}'s message: {:?}.", redactor, original_sender, reason)
-        }
-        Some((redactor, None)) => {
-            if redactor == event_tl_item.sender() {
-                format!("{} deleted their own message.", original_sender)
-            } else {
-                format!("{} deleted {}'s message.", redactor, original_sender)
-            }
-        }
-        None => {
-            format!("{}'s message was deleted.", original_sender)
-        }
     }
 }
 
@@ -2334,8 +2170,8 @@ impl SmallStateEventContent for timeline::OtherState {
         _item_drawn_status: ItemDrawnStatus,
         mut new_drawn_status: ItemDrawnStatus,
     ) -> (WidgetRef, ItemDrawnStatus) {
-        let item = if let Some(text) = preview_text_of_other_state(self) {
-            item.label(id!(content)).set_text(&format!("{username} {text}"));
+        let item = if let Some(text_preview) = text_preview_of_other_state(self) {
+            item.label(id!(content)).set_text(&text_preview.format_with(username));
             new_drawn_status.content_drawn = true;
             item
         } else {
@@ -2344,76 +2180,6 @@ impl SmallStateEventContent for timeline::OtherState {
             item
         };
         (item, new_drawn_status)
-    }
-}
-
-pub fn preview_text_of_other_state(
-    other_state: &timeline::OtherState,
-) -> Option<String> {
-    match other_state.content() {
-        AnyOtherFullStateEventContent::RoomAliases(FullStateEventContent::Original { content, .. }) => {
-            let mut s = format!("set this room's aliases to ");
-            let last_alias = content.aliases.len() - 1;
-            for (i, alias) in content.aliases.iter().enumerate() {
-                s.push_str(alias.as_str());
-                if i != last_alias {
-                    s.push_str(", ");
-                }
-            }
-            s.push_str(".");
-            Some(s)
-        }
-        AnyOtherFullStateEventContent::RoomAvatar(_) => {
-            Some(format!("set this room's avatar picture."))
-        }
-        AnyOtherFullStateEventContent::RoomCanonicalAlias(FullStateEventContent::Original { content, .. }) => {
-            Some(format!("set the main address of this room to {}.",
-                content.alias.as_ref().map(|a| a.as_str()).unwrap_or("none")
-            ))
-        }
-        AnyOtherFullStateEventContent::RoomCreate(FullStateEventContent::Original { content, .. }) => {
-            Some(format!("created this room (v{}).", content.room_version.as_str()))
-        }
-        AnyOtherFullStateEventContent::RoomGuestAccess(FullStateEventContent::Original { content, .. }) => {
-            Some(match content.guest_access {
-                GuestAccess::CanJoin => format!("has allowed guests to join this room."),
-                GuestAccess::Forbidden | _ => format!("has forbidden guests from joining this room."),
-            })
-        }
-        AnyOtherFullStateEventContent::RoomHistoryVisibility(FullStateEventContent::Original { content, .. }) => {
-            let visibility = match content.history_visibility {
-                HistoryVisibility::Invited => "invited users, since they were invited.",
-                HistoryVisibility::Joined => "joined users, since they joined.",
-                HistoryVisibility::Shared => "joined users, for all of time.",
-                HistoryVisibility::WorldReadable | _ => "anyone for all time.",
-            };
-            Some(format!("set this room's history to be visible by {}.", visibility))
-        }
-        AnyOtherFullStateEventContent::RoomJoinRules(FullStateEventContent::Original { content, .. }) => {
-            Some(match content.join_rule {
-                JoinRule::Public => format!("set this room to be joinable by anyone."),
-                JoinRule::Knock => format!("set this room to be joinable by invite only or by request."),
-                JoinRule::Private => format!("set this room to be private."),
-                JoinRule::Restricted(_) => format!("set this room to be joinable by invite only or with restrictions."),
-                JoinRule::KnockRestricted(_) => format!("set this room to be joinable by invite only or requestable with restrictions."),
-                JoinRule::Invite | _ => format!("set this room to be joinable by invite only."),
-            })
-        }
-        AnyOtherFullStateEventContent::RoomName(FullStateEventContent::Original { content, .. }) => {
-            Some(format!("changed this room's name to {:?}.", content.name))
-        }
-        AnyOtherFullStateEventContent::RoomPowerLevels(_) => {
-            None
-        }
-        AnyOtherFullStateEventContent::RoomTopic(FullStateEventContent::Original { content, .. }) => {
-            Some(format!("changed this room's topic to {:?}.", content.topic))
-        }
-        AnyOtherFullStateEventContent::SpaceParent(_)
-        | AnyOtherFullStateEventContent::SpaceChild(_) => None,
-        _other => {
-            // log!("*** Unhandled: {:?}.", _other);
-            None
-        }
     }
 }
 
@@ -2430,40 +2196,14 @@ impl SmallStateEventContent for MemberProfileChange {
         mut new_drawn_status: ItemDrawnStatus,
     ) -> (WidgetRef, ItemDrawnStatus) {
         item.label(id!(content)).set_text(
-            &preview_text_of_member_profile_change(self, username)
+            &text_preview_of_member_profile_change(self, username)
+                .format_with(username)
         );
         new_drawn_status.content_drawn = true;
         (item, new_drawn_status)
     }
 }
 
-pub fn preview_text_of_member_profile_change(
-    change: &MemberProfileChange,
-    username: &str,
-) -> String {
-    let name_text = if let Some(name_change) = change.displayname_change() {
-        let old = name_change.old.as_deref().unwrap_or(&username);
-        if let Some(new) = name_change.new.as_ref() {
-            format!("{old} changed their display name to {new:?}")
-        } else {
-            format!("{old} removed their display name")
-        }
-    } else {
-        String::new()
-    };
-
-    let avatar_text = if let Some(_avatar_change) = change.avatar_url_change() {
-        if name_text.is_empty() {
-            format!("{} changed their profile picture", username)
-        } else {
-            format!(" and changed their profile picture")
-        }
-    } else {
-        String::new()
-    };
-
-    format!("{}{}.", name_text, avatar_text)
-}
 
 impl SmallStateEventContent for RoomMembershipChange {
     fn populate_item_content(
@@ -2477,7 +2217,7 @@ impl SmallStateEventContent for RoomMembershipChange {
         _item_drawn_status: ItemDrawnStatus,
         mut new_drawn_status: ItemDrawnStatus,
     ) -> (WidgetRef, ItemDrawnStatus) {
-        let Some(text) = preview_text_of_room_membership_change(self) else {
+        let Some(preview) = text_preview_of_room_membership_change(self) else {
             // Don't actually display anything for nonexistent/unimportant membership changes.
             return (
                 list.item(cx, item_id, live_id!(Empty)).unwrap(),
@@ -2485,57 +2225,12 @@ impl SmallStateEventContent for RoomMembershipChange {
             );
         };
 
-        item.label(id!(content)).set_text(&format!("{username} {text}"));
+        item.label(id!(content)).set_text(&preview.format_with(username));
         new_drawn_status.content_drawn = true;
         (item, new_drawn_status)
     }
 }
 
-pub fn preview_text_of_room_membership_change(
-    change: &RoomMembershipChange,
-) -> Option<String> {
-    let dn = change.display_name();
-    let change_user_id = dn.as_deref()
-        .unwrap_or_else(|| change.user_id().as_str());
-    let text = match change.change() {
-        None
-        | Some(MembershipChange::NotImplemented)
-        | Some(MembershipChange::None)
-        | Some(MembershipChange::Error) => {
-            // Don't actually display anything for nonexistent/unimportant membership changes.
-            return None;
-        }
-        Some(MembershipChange::Joined) =>
-            format!("joined this room."),
-        Some(MembershipChange::Left) =>
-            format!("left this room."),
-        Some(MembershipChange::Banned) =>
-            format!("banned {} from this room.", change_user_id),
-        Some(MembershipChange::Unbanned) =>
-            format!("unbanned {} from this room.", change_user_id),
-        Some(MembershipChange::Kicked) =>
-            format!("kicked {} from this room.", change_user_id),
-        Some(MembershipChange::Invited) =>
-            format!("invited {} to this room.", change_user_id),
-        Some(MembershipChange::KickedAndBanned) =>
-            format!("kicked and banned {} from this room.", change_user_id),
-        Some(MembershipChange::InvitationAccepted) =>
-            format!("accepted an invitation to this room."),
-        Some(MembershipChange::InvitationRejected) =>
-            format!("rejected an invitation to this room."),
-        Some(MembershipChange::InvitationRevoked) =>
-            format!("revoked {}'s invitation to this room.", change_user_id),
-        Some(MembershipChange::Knocked) =>
-            format!("requested to join this room."),
-        Some(MembershipChange::KnockAccepted) =>
-            format!("accepted {}'s request to join this room.", change_user_id),
-        Some(MembershipChange::KnockRetracted) =>
-            format!("retracted their request to join this room."),
-        Some(MembershipChange::KnockDenied) =>
-            format!("denied {}'s request to join this room.", change_user_id),
-    };
-    Some(text)
-}
 
 /// Creates, populates, and adds a SmallStateEvent liveview widget to the given `PortalList`
 /// with the given `item_id`.
